+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
-### Added
+### Removed
 
-- Occupied room counter for twin room 
+- Occupied room counters for single and double rooms with reset buttons
 
 ## 0.1.0 - 2023-12-07